chore: move analytics related methods from CH reader to their own mod… (#5935)
